[src] Move positional patchembedding to vit.py
