+++ audyn/models/ssast.py
@@ -13,16 +13,10 @@ import torch.nn.functional as F
 from omegaconf import OmegaConf
 from torch.nn.common_types import _size_2_t
 
+from ..modules.vit import PositionalPatchEmbedding
 from ..utils import instantiate, model_cache_dir
 from ..utils.github import download_file_from_github_release
-from .ast import (
-    Aggregator,
-    BaseAudioSpectrogramTransformer,
-    Head,
-    MLPHead,
-    PositionalPatchEmbedding,
-    _align_patch_embedding,
-)
+from .ast import Aggregator, BaseAudioSpectrogramTransformer, Head, MLPHead, _align_patch_embedding
 
 __all__ = [
     "SelfSupervisedAudioSpectrogramTransformerMaskedPatchModel",
